:truck: move `Course` to `src/evaluation` context
