+++ api/tests/shared/unit/domain/models/Course_test.js
@@ -1,4 +1,4 @@
-import { Course } from '../../../../../src/shared/domain/models/Course.js';
+import { Course } from '../../../../../src/evaluation/domain/models/Course.js';
 import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Domain | Models | Course', function () {
