+++ api/tests/shared/unit/infrastructure/serializers/jsonapi/course-serializer_test.js
@@ -1,4 +1,4 @@
-import { Course } from '../../../../../../src/shared/domain/models/Course.js';
+import { Course } from '../../../../../../src/evaluation/domain/models/Course.js';
 import * as serializer from '../../../../../../src/shared/infrastructure/serializers/jsonapi/course-serializer.js';
 import { expect } from '../../../../../test-helper.js';
