+++ api/tests/tooling/domain-builder/factory/build-course.js
@@ -1,4 +1,4 @@
-import { Course } from '../../../../src/shared/domain/models/Course.js';
+import { Course } from '../../../../src/evaluation/domain/models/Course.js';
 
 const buildCourse = function ({
   id = 'recCOUR123',
